마스크 포장 탭메뉴 수정, 생산랏 업데이트 시간 컬럼 추가

+++ src/main/java/com/busience/standard/dto/PaldangPackagingStandardDto.java
@@ -14,7 +14,7 @@ public class PaldangPackagingStandardDto {
 	private String packaging_Type;
 	private String packaging_Size;
 	private String packaging_Item;
-	private double packaging_Small;
-	private double packaging_Large;
+	private int packaging_Small;
+	private int packaging_Large;
 	private String packaging_ModifyDate;
 }
